core_services/apps/company_admin/serializers/temporary_password_reset_serializer.py
from rest_framework import serializers


class TemporaryPasswordResetSerializer(serializers.Serializer):
    temporary_password = serializers.CharField(required=True, allow_blank=False, min_length=8, max_length=128)
