core_services/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-yasg==1.21.11
python-decouple==3.8
psycopg2-binary==2.9.11
django-redis==6.0.0
celery==5.5.3
redis==6.4.0
pymongo==4.10.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
argon2-cffi==25.1.0
razorpay==2.0.0
Pillow==11.3.0
reportlab==4.2.2
django-prometheus==2.4.0
boto3==1.39.11
django-storages==1.14.6
firebase-admin==7.1.0
pytest==8.4.1
pytest-django==4.11.1
